refactor: extract discard path into dev/discard.rs
Move discard and __discard_one_cluster into a new dev/discard.rs
submodule as an impl extension of Qcow2Dev. No visibility changes
needed. No logic changes.
